Removed unused hangman files
